% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDSSPdata.R
\name{addDSSPdata}
\alias{addDSSPdata}
\title{Searches the secondary structure of multiple aminoacids in .dssp files}
\usage{
addDSSPdata(aantinfo, ntinfo, PATH = "./")
}
\arguments{
\item{aantinfo}{is the data.frame output of make_aantinfo function}

\item{ntinfo}{is the output of make_ntinfo function (it can be replaced by a 
data.frame with the columns ntID and pdbID)}

\item{PATH}{is the directory where DSSP has been previously executed. The
output files of DSSP should be named "pdbID".dssp}
}
\value{
an updated version of aantinfo with the secondary structure data for
   the amino acids in the data.frame
}
\description{
Given the data.frame "aantinfo" and a path with the dssp files, it will
read the file and extract the info about the secondary structure of the
aminoacid in contact with each of the nucleotides.
}
\author{
Diego Gallego
}
