% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{angles_dist}
\alias{angles_dist}
\title{Plot distribution of desired angles in circular plots}
\usage{
angles_dist(ntID = NULL, ntinfo, angles = c("alpha", "beta", "gamma",
  "delta", "epsilon", "zeta", "chi", "pu_phase"), o = "dihedrals",
  width = 15, height = 15, bg = "white", units = "cm", res = 200,
  cex = 0.6, cols = 3)
}
\arguments{
\item{ntID}{An obejct of class vector with the desired nucleotides of 
analysis. If NULL all the nucleotides in the data.frame will be used}

\item{ntinfo}{A data.frame with the input data. It should contain the 
columns with the desired angles and a column labeled ntID}

\item{angles}{The column names with the desired data}

\item{o}{the name of the output file (the function will automatically 
include the .png extension)}

\item{width}{The width of the plot (passed to the png() function)}

\item{height}{The height of the plot (passed to the png() function)}

\item{bg}{The background color of the plot (passed to the png() function)}

\item{units}{The unit to measure height and width (passed to the png() 
function)}

\item{res}{Resolution (passed to the png() function)}

\item{cex}{To be passed to the par() function}

\item{cols}{Number of columns in the ouput picture.}
}
\value{
A ".png" file with the desired plots
}
\description{
Given a data.frame with nucleotides data, it generates a series of circular
plots with the desired angles. NA in the data are ignored.
}
\author{
Diego Gallego
}
