% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyToPDB.R
\name{applyToPDB}
\alias{applyToPDB}
\title{Applies a function over a list of PDB ID entries.}
\usage{
applyToPDB(FUNCTION, listpdb = NULL, verbose = T, as.df = T, ...)
}
\arguments{
\item{FUNCTION}{A function of interest}

\item{listpdb}{A list/vector containing the PDB IDs of interest. If NULL,
the complete list of PDB entries is downloaded and used.}

\item{verbose}{A logical to print details of the loop process}

\item{as.df}{A logical that stands for "as.data.frame". If TRUE, the output
will be returned in the form of a data.frame, otherwise a list.}

\item{...}{optional arguments to ‘FUNCTION’}
}
\value{
A data.frame with the PDB IDs (first colunm) and the output of the
function of interest (second column) or a list with the results.
}
\description{
Given a function of interest, it is applied to all the PDB entries.
See supported functions in ?query_functions.
}
\examples{
listpdb <- c("1s72", "1bau", "1rna")
applyToPDB(query_technique, listpdb, verbose=F)

}
\author{
Diego Gallego
}
