% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BasicCIF-class}
\alias{BasicCIF-class}
\alias{BasicCIF}
\title{An S4 class to represent the mmCIF file.}
\description{
An S4 class to represent the mmCIF file.
}
\section{Slots}{

\describe{
\item{\code{entry}}{The ID code}
}}

