% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bp_hist}
\alias{bp_hist}
\title{Plot histogram}
\usage{
bp_hist(ntID, ntinfo, categories = "LW", o = "bp", main = o, rm.na = F,
  width = 15, height = 15, bg = "white", units = "cm", res = 200,
  cex = 0.5)
}
\description{
Plot histogram
}
\author{
Diego Gallego
}
