% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CIF-class}
\alias{CIF-class}
\alias{CIF}
\title{An S4 class to represent a structure parsed from its mmCIF file.}
\description{
An S4 class to represent a structure parsed from its mmCIF file.
}
\section{Slots}{

\describe{
\item{\code{entry}}{The ID code}

\item{\code{audit_conform}}{'mmCIF' dictionary version}

\item{\code{database_2}}{Cross-reference ID codes for other databases}

\item{\code{pdbx_database_status}}{Deposition data}

\item{\code{audit_author}}{Author data}

\item{\code{entity}}{Entities (molecules & ions) in the structure}

\item{\code{chem_comp}}{Residues (ATOM & HETATM) in the structure}

\item{\code{exptl}}{Experimental technique}

\item{\code{struct}}{Author description of the structure}

\item{\code{struct_keywords}}{Author description key words}

\item{\code{struct_asym}}{Chain-entity equivalences}

\item{\code{atom_sites}}{Details about the crystallographic cell}

\item{\code{atom_type}}{Details about the atoms in structure}

\item{\code{atom_site}}{The atomic coordinates}
}}

\seealso{
To create CIF objects use [parse.cif()]
}
\author{
Diego Gallego
}
