% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CIF.R
\docType{methods}
\name{cifAsPDB}
\alias{cifAsPDB}
\alias{cifAsPDB,CIF-method}
\alias{cifAsPDB,character-method}
\title{Coerce CIF S4 object to pdb S3 object as found in bio3d package}
\usage{
cifAsPDB(cif, model = NULL, chain = NULL, alt = c("A"))

\S4method{cifAsPDB}{CIF}(cif, model = NULL, chain = NULL, alt = c("A"))

\S4method{cifAsPDB}{character}(cif, model = NULL, chain = NULL,
  alt = c("A"))
}
\arguments{
\item{cif}{A CIF object as obtained from cifParser. It can also accept
a 4-character PDB ID.}

\item{model}{A string with the model number (in case you are only
interested in a particular model) If NULL, all models are read. This option
will only be applied to the "atom" attribute of the output pdb object.}

\item{chain}{A string with the chain identifier (in case you are only
interested in a particular chain). If NULL, all chains are read. This option
will only be applied to the "atom" attribute of the output pdb object.}

\item{alt}{A string or a vector of strings with the desired alternative
records. This option will only be applied to the "atom" attribute of the
output pdb object.}
}
\value{
A pdb object compatible with bio3d (Grant et al. 2006) functions.
}
\description{
Coerces CIF to pdb class
}
\examples{
cif <- cifParser("1bau")
pdb <- cifAsPDB(cif)

}
\author{
Diego Gallego
}
