% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CIF.R
\name{cifDownload}
\alias{cifDownload}
\title{Download Protein Data Bank structures}
\usage{
cifDownload(pdbID, destfile = NULL, extension = ".cif.gz",
  URL = NULL, verbose = FALSE)
}
\arguments{
\item{pdbID}{A 4 character string that matches a structure in the Protein
Data Bank.}

\item{destfile}{File name to save the downloaded structure. If NULL, a
temporal directory is used and  the file name is constructed based on
the PDB ID and extension.}

\item{extension}{A string with the desired file extension.}

\item{URL}{A string with the URL of use. If NULL, the RCSB is used as
default.}

\item{verbose}{A logical to print process details.}
}
\value{
The file name.
}
\description{
Given a 4-character string (PDB ID), download structure.
}
\examples{
    ## cifDownload("1bau")

}
\author{
Diego Gallego
}
