% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CIF.R
\name{.cifMakeSure}
\alias{.cifMakeSure}
\title{Is it a CIF object? Make it be!}
\usage{
.cifMakeSure(cif, verbose = F, check = "cifCheck")
}
\arguments{
\item{cif}{A cif object obtained from cifParser or a pdb ID so that the
function can download the data.}

\item{verbose}{A logical indicating whether to print details of the process}

\item{check}{A string with the name of the function to use. It has been
thought to be used with 'cifCheck' function.}
}
\value{
A cif object, which might be the same input or the downloaded data
}
\description{
Internal function to check if an input is actually a CIF object
If not, the cif file is read from the MMB API.
}
\author{
Diego Gallego
}
