% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CIF.R
\docType{methods}
\name{cifParser}
\alias{cifParser}
\alias{cifParser,ANY-method}
\title{Parse coordinates from CIF files}
\usage{
cifParser(pdbID, verbose = FALSE)

\S4method{cifParser}{ANY}(pdbID, verbose = FALSE)
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure in the Protein
Data Bank (or an existing file in disk).}

\item{verbose}{A logical indicating whether to print details of the process.}
}
\value{
A S4 CIF object.
}
\description{
Given a file or PDB ID, the function parses the coordinates of the
structure. It can also read all the fields of the mmCIF format.
}
\examples{
    cif <- cifParser("1bau")

}
\author{
Diego Gallego
}
