% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyRNA.R
\name{classifyRNA}
\alias{classifyRNA}
\title{Classify RNA structures}
\usage{
classifyRNA(pdbID, length = 3, ...)
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure ID in the
Protein Data Bank.}

\item{length}{A positive integer to use as a threshold to classify RNA in
the NoRNA group.}

\item{...}{Arguments to be passed to query function (see ?queryFunctions).}
}
\value{
A string with the type of RNA.
}
\description{
From different queries to databases, the function classifies a structure in
different groups:\cr
\itemize{
\item {NoRNA}: the structure does not contain RNA or it is shorter than a
threshold set by "length".\cr
\item {nakedRNA}: the only molecule(s) in the PDB ID is RNA.\cr
\item {protRNA}: the PDB contains a protein-RNA complex.\cr
\item {DprotRNA}: the PDB contains a protein-RNA complex and the protein has D
aminoacids.\cr
\item {DNARNA}: the PDB contains a DNA-RNA hybrid.\cr
\item {PNARNA}: the PDB contains a PNA-RNA hybrid.\cr
\item {ANARNA}: the PDB contains a ANA-RNA hybrid.\cr
\item {LNARNA}: the PDB contains a LNA-RNA hybrid.\cr
\item {ligandRNA}: the RNA is interacting with an organic ligand, ions are not
considered.\cr
}
}
\examples{
classifyRNA("1S72")

}
\author{
Diego Gallego
}
