% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_entities.R
\name{distances_entities}
\alias{distances_entities}
\title{Computes distances between all the atoms of selected entities in a mmCIF 
structure}
\usage{
distances_entities(cif, model = NULL, refent, entities = c("all"), n = 1,
  cutoff = c(0, 5), verbose = T, detailedoutput = T,
  data_of_interest = NULL)
}
\arguments{
\item{cif}{A cif object obtained from parse.cif or a pdb ID.}

\item{model}{The model of interest to use in the calculations. The first 
model is always the default.}

\item{refent}{A string with the entity ID of reference. The distance output
will be referred to the atoms/residues of this entity.}

\item{entities}{A character vector with the entities of interest. The 
default "all" will select all of them except the refent.}

\item{n}{An integer indicating how many atoms to return. The default 
n=1 returns only the minimum distance, the closest atom; n=2 would return
the two closest atoms and so on. In all cases, only atoms within the 
cutoff will be returned.}

\item{cutoff}{A numeric indicating the maximum radius to consider in 
angstroms. Atoms further than the cutoff won't be returned.}

\item{verbose}{A logical indicating whether to print details of the process.}

\item{detailedoutput}{A logical indicating whether to include additional 
information for each atom (see data_of_interest below). If FALSE, only
the eleno (element number) are returned.}

\item{data_of_interest}{A vector of strings. Only used if detailedoutput is 
TRUE. The vector should only contain the strings between the following:
"type", "elety", "alt", "resid", "chain", "resno", "insert", "x", "y",
"z", "o", "b", "entid", "elesy", "charge", "asym_id", "seq_id", 
"comp_id", "atom_id", "model".
The selected fields will be returned for both atoms.}
}
\value{
A data.frame with the nearest neighbour information.
}
\description{
Given a cif object (or a pdb ID), the function computes the distances 
between atoms of the selected entity IDs. For each atom/residue of the 
reference entity the function returns the closest atoms of the other 
entities. 
To see the entities and their IDs of a given structure run:
   cif <- parse.cif("XXXX") #Where XXXX is the pdb ID (e.g.1S72)
   View(cif$entity)
}
\author{
Diego Gallego
}
