% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities-data.R
\docType{data}
\name{entities}
\alias{entities}
\title{List of NA containing PDB IDs and related data}
\format{An object of class data.frame}
\usage{
data(entities)
}
\value{
data.frame with a list and features of NA PDB IDs
}
\description{
Number of each type of entity in a set of Nucleic Acid containing
structures.
}
\keyword{datasets}
