% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastquery-data.R
\docType{data}
\name{fastquery}
\alias{fastquery}
\title{List of NA containing PDB IDs and related data}
\format{An object of class data.frame}
\usage{
data(fastquery)
}
\value{
data.frame with a list and features of NA PDB IDs
}
\description{
Data from particularly slow queries.
}
\keyword{datasets}
