% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alternative_representants.R
\name{get_alternative_representants}
\alias{get_alternative_representants}
\title{Get Alternative Representants}
\usage{
get_alternative_representants(data, technique = NULL, resol = NULL,
  type = NULL, verbose = F)
}
\arguments{
\item{data}{The output of get_leontis_nr_list.}

\item{technique}{One or more techniques of interest (For correct use, see
example below). For the list of techniques, see "bio3dRNA:::allowedtechs".}

\item{resol}{A positive real number to specify a desired resolution.}

\item{type}{A string indicating the type of desired RNA, according with
the RNAclassifier function.}
}
\value{
A data.frame with info about all the "Equivalence Classes" and
the selected Representants according with the specified conditions.
}
\description{
This function is closely related with get_leontis_nr_list(). From its
output, the family members of each equivalence class are checked for
desired features. The first member that matches all the desired features
is returned.
Note: The function might seem slow (really slow sometimes), but it is
doing lots of queries under the hood. Set verbose to TRUE to see evolution.
The example below works with a small dataset and takes ~1 min.
}
\examples{
 data <- get_leontis_nr_list( release = 3.2, threshold = "1.5A" )
 alternative <- get_alternative_representants( data = data, 
         type = "nakedRNA")

}
\author{
Diego Gallego
}
