% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leontis_nr_list.R
\name{get_leontis_nr_list}
\alias{get_leontis_nr_list}
\title{Download Representative List of RNA structures}
\usage{
get_leontis_nr_list(release, threshold = "all")
}
\arguments{
\item{release}{A number indicating the list of interest.}

\item{threshold}{A string that matches one of the lists in the BGSU RNA 
site ("1.5A", "2.0A", "2.5A", "3.0A", "3.5A", "4.0A", "20.0A" or "all").
Note that "all" returns structures solved by all the possible techniques.}
}
\value{
A data frame with the list of Equivalence Classes and the Representant
  and Members of each Equivalence Class. Note that the output is formated 
  according with Leontis&Zirbel nomenclature ( AAAA|M|C ), where "AAAA" is
  the PDB accession code, "M" is the model and "C" is the Chain to be used.
}
\description{
According with Leontis & Zirbel, 2012 (Nonredundant 3D Structure Datasets 
for RNA Knowledge Extraction and Benchmarking), the PDB contains structures
that are redundant. Their work resulted in the BGSU RNA Site 
(http://rna.bgsu.edu/rna3dhub/nrlist/) where one can find weekly releases 
of Representative Sets of RNA structures (formerly called non-redundant 
lists). This function access their website and returns the desired list.
}
\author{
Diego Gallego
}
