% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_x_hetAtms.R
\name{has_x_hetAtms}
\alias{has_x_hetAtms}
\title{Check if a given PDB contains the ligand/modbase of interest}
\usage{
has_x_hetAtms(pdbID, hetAtms)
}
\arguments{
\item{pdbID}{A 4-character string.}

\item{hetAtms}{A string with the ligand/modbase ID.}
}
\value{
A logical. TRUE if the given hetAtms is present in the structure.
}
\description{
Given a 4-character string (PDB ID) and a ligand/modbase ID, the MMB API is
accessed to check the presence of the ligand/modres in the given PDB ID.
}
\author{
Diego Gallego
}
