% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aantinfo.R
\name{make_aantinfo}
\alias{make_aantinfo}
\title{Obtain nucleotide-protein interactions from a data set of RNA structures}
\usage{
make_aantinfo(pdbID, model = NULL, chain = NULL, ntinfo = NULL,
  path = NULL, extension = NULL, cores = 1, cutoff = 15, ...)
}
\arguments{
\item{pdbID}{A list/vector containing the desired PDB IDs or a list of pdb
objects as provided by "read.pdb", "read.cif", "cifParser" ...}

\item{model}{A vector with same length of pdbID containing the
desired model for each pdbID. If all models are desired, use "all".
If no models are specified, the first one will be used for each pdbID}

\item{chain}{A vector with same length of pdbID containing the
desired chain for each pdbID. If no chain is specified, all chains will
be analysed by default. Non-nucleic acid chains will be ignored.}

\item{path}{Directory in which the PDB/CIF files can be found (if NULL, the
function will download them). If you provide a "path", make sure the
file names are the PDB IDs followed by ".cif" or "pdb". The function
will find them using the strings in pdbID, so make sure you use the
same case.}

\item{extension}{A string matching the files extension (e.g. ".pdb",
".cif", "pdb.gz", "cif.gz").
Only necessary if the PDB files are to be read from disk and a path is
provided.}

\item{cores}{Number of CPU cores to be used.}

\item{...}{Arguments to be passed to \code{\link[=measureNuc]{measureNuc()}}}

\item{range}{A numeric vector with two values to indicate the desired
length range for the Nucleic Acid chains. If a chain falls outside the
range, it is not analysed.}
}
\value{
A data.frame with data about every nucleotide in the input set
}
\details{
Generates a data.frame with the data about the closests ribonucleotides to
the protein for a list of PDB. The data can be related to unique nucleotide
indentifiers (ntID) by providing the output of \code{\link[=getNucData]{getNucData()}}.
}
\examples{
 ## This is a toy example, see vignettes for more usages.
 pdblist <- list("1bau", "2rn1")
 model <- list("1", "2")
 chain <- list("all", "all")
 ntinfo <- getNucData(pdbID=pdblist, model=model, chain=chain)

}
\author{
Diego Gallego
}
