% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sure_is_cif.R
\name{make_sure_is_cif}
\alias{make_sure_is_cif}
\title{Is a cif object? Make it be!}
\usage{
make_sure_is_cif(cif, verbose = F, check = "is.cif")
}
\arguments{
\item{cif}{A cif object obtained from parse.cif or a pdb ID so that the
function can download the data.}

\item{verbose}{A logical indicating whether to print details of the process}

\item{check}{A string with the name of the function to use. It has been
thought to be used with 'is.cif' or 'is.pdb' functions.}
}
\value{
A cif object, which might be the same input or the downloaded data
}
\description{
Internal function to check if a cif/pdb input is actually a cif/pdb object
If not, the cif file is read from the MMB API.
}
\author{
Diego Gallego
}
