% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.select.R
\name{model.select}
\alias{model.select}
\title{Selects a desired model from a PDB structure}
\usage{
model.select(pdb, model, verbose = T)
}
\arguments{
\item{pdb}{A pdb object with multiple models (obtained from read.cif.RAM
or read.cif/read.pdb).}

\item{model}{A string with the desired model number.}

\item{verbose}{A logical to print messages on screen}
}
\value{
A pdb object with the desired model coordinates.
}
\description{
Given a pdb object read using read.cif.RAM function OR read.cif/read.pdb
functions (from bio3d) specifying "multi = TRUE", the function returns a
pdb object with the desired model. 
Since some structures deposited in the PDB contain models with different
number of atoms, the pdb objects in R require a special treatment. 
The read.cif.RAM and model.select functions can cope with these structures
(e.g. 1JTW).
}
\author{
Diego Gallego
}
