% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSetOfDistributions}
\alias{plotSetOfDistributions}
\title{Plot distribution of desired angles in circular plots}
\usage{
plotSetOfDistributions(ntinfo, ntID = NULL, angles = c("alpha", "beta",
  "gamma", "delta", "epsilon", "zeta", "chi", "pu_phase"), cex = 0.6,
  cols = 3, file = NULL, width = 15, height = 15, bg = "white",
  units = "cm", res = 200)
}
\arguments{
\item{ntinfo}{A data.frame with the input data. It should contain the
columns with the desired angles and a column labeled ntID.}

\item{ntID}{A vector of integers with the desired nucleotides of
analysis. If NULL all the nucleotides in the data.frame will be used.}

\item{angles}{The column names with the desired data.}

\item{cex}{To be passed to the \code{par}.}

\item{cols}{Number of columns in the ouput picture.}

\item{file}{A string with the name of the output file. If NULL, the
plot will be printed to screen.}

\item{width}{The width of the plot to be passed to \code{png}.}

\item{height}{The height of the plot to be passed to \code{png}.}

\item{bg}{The background color of the plot to be passed to \code{png}.}

\item{units}{The unit to measure height and width to be passed
to \code{png}.}

\item{res}{Resolution to be passed to \code{png}.}
}
\value{
A series of circular plots with the distributions of the desired
angles, which can be directly saved  to a ".png" file.
}
\description{
Given a data.frame with nucleotides data, it generates a series of circular
plots for the desired angles. NA in the data are ignored.
}
\examples{
    #ntinfo <- pipeNucData("1bau")
    #C3endo_ntID <- cleanByPucker(ntinfo, pucker="C3'endo")

    ## Plot torsional angles for C3'endo nucleotides
    #plotSetOfDistributions(ntinfo=ntinfo, ntID=C3endo_ntID, 
    #                        file="1bau_C3endo.png")

    ## Which is the same as doing:
    #C3endo <- ntinfo[ntinfo$ntID \%in\% C3endo_ntID,]
    #plotSetOfDistributions(ntinfo=C3endo, file="1bau_C3endo.png")

}
\author{
Diego Gallego
}
