% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_functions}
\alias{query_functions}
\alias{query_technique}
\alias{query_resol}
\alias{query_compound}
\alias{query_header}
\alias{query_compType}
\alias{query_NDBId}
\alias{query_authors}
\alias{query_depdate}
\alias{query_reldate}
\alias{query_revdate}
\alias{query_formats}
\alias{query_entities}
\alias{query_hetAtms}
\alias{query_chains}
\alias{query_modres}
\alias{query_orgligands}
\title{General functions to query PDB data from APIs}
\usage{
query_technique(pdbID, ...)

query_resol(pdbID, ...)

query_compound(pdbID, ...)

query_header(pdbID, ...)

query_compType(pdbID, ...)

query_NDBId(pdbID, ...)

query_authors(pdbID, ...)

query_depdate(pdbID, ...)

query_reldate(pdbID, ...)

query_revdate(pdbID, ...)

query_formats(pdbID, ...)

query_entities(pdbID, ...)

query_hetAtms(pdbID, NAtoNa = T, ...)

query_chains(pdbID, chain = NULL, subset = NULL, ...)

query_modres(pdbID, onlymodres = F, ...)

query_orgligands(pdbID, ...)
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure in the Protein
Data Bank.}

\item{...}{For advanced usage, arguments to be passed to subfunction
"queryAPI". Type ?queryAPI for documentation.}

\item{NAtoNa}{A logical. If TRUE, sodium ion (NA) is modified as "Na".}

\item{chain}{A string with the chain identifier (in case you are only
interested in a particular chain). If NULL, the info about all the chains
is returned.}

\item{subset}{Optional argument indicating "type", "length" or
"description". If NULL, all the columns in the data.frame are returned.}

\item{onlymodres}{A logical. If TRUE, only the modified residues are
returned.}
}
\value{
A vector with the desired information retrieved from the API.
}
\description{
Suported APIs: MMB (Molecular Modeling and Bioinformatics Laboratoy,
Dr.Orozco & Dr.Gelpí groups) and
EBI (The European Bioinformatics Institute, EMBL-EBI). Take into account
that due to the different structure of the APIs, the outputs might be
formatted differently.\cr
Most functions are self-explanatory, but some details might be of interest:
\cr
\itemize{
\item query_resol: Keep in mind that the resolution will only be returned
if the structure was solved by X-RAY or Cryo-EM.\cr
\item query_compound: It returns the title as it appears in the header (PDB
format) or in the "_struct.title" field (mmCIF format).\cr
\item query_header: It returns the classification of the structure as it
appears in the header (PDB format) or in the
"_struct_keywords.pdbx_keywords" field (mmCIF format). (Feature only
accessible through the MMB API).\cr
\item query_compType: It returns the type as defined in the PDB
(e.g. prot-nuc). (Feature only accessible through the MMB API).\cr
\item query_NDBId: It returns the ID of the same structure in the Nucleic
Acid Database. (Feature only accessible through the MMB API).\cr
\item query_depdate: It returns the deposition date.\cr
\item query_reldate: It returns the release date. (Feature only accessible
through the EBI API).\cr
\item query_revdate: It returns the revision date. (Feature only accessible
through the EBI API).\cr
\item query_formats: It returns a vector with the available files (e.g. pdb).
}
}
\author{
Diego Gallego
}
