% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_pdblist.R
\name{query_pdblist}
\alias{query_pdblist}
\title{Downloads the list of ID of ALL current PDB entries}
\usage{
query_pdblist()
}
\value{
A vector with all the PDB ID entries (updated weekly).
}
\description{
Function to get the list of PDB IDs in the Protein Data Bank.
}
\author{
Diego Gallego
}
