% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAPI.R
\name{queryAPI}
\alias{queryAPI}
\title{Launch queries to the MMB or EBI APIs}
\usage{
queryAPI(pdbID, info, API = "default", reuse = F, envir = parent.frame(n =
  2), verbose = F)
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure in the Protein
Data Bank.}

\item{info}{A string with the desired query nickname.}

\item{API}{A string that matches "mmb" or "ebi".}

\item{reuse}{A logical. Set to TRUE if the same query is going to be send
repeatedly, so that the result is saved in RAM (ir provides faster
user access and avoids unnecessary work in the servers).}

\item{envir}{Environment to save&retrieve the data if reuse is TRUE.}

\item{verbose}{A logical. TRUE to print details of the process.}
}
\value{
A vector or data.frame with the desired data.
}
\description{
Given a 4-character string (PDB ID) and the desired "info", it sends  a
query to the desired API and returns the output. This is a wrapper
function called by most of the query_functions (for documentation see
?query_functions).
}
\author{
Diego Gallego
}
