% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.cif.RAM.R
\name{read.cif.RAM}
\alias{read.cif.RAM}
\title{Parse coordinates from CIF files}
\usage{
read.cif.RAM(pdbID, model = NULL, chain = NULL, alt = c("A"))
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure in the Protein 
Data Bank.}

\item{model}{A string with the model number (in case you are only
interested in a particular model) If NULL, all models are read.}

\item{chain}{A string with the chain identifier (in case you are only
interested in a particular chain). If NULL, all chains are read.}

\item{alt}{A string or a vector of strings with the desired alternative
records.}
}
\value{
A "pdb" object as defined in the bio3d package (Grant et al. 2006)
}
\description{
Given a PDB ID, the function access the MMB API and downloads directly to
RAM the coordinates of the desired structure. It can also read CIF files in
Linux systems.
}
\author{
Diego Gallego
}
