% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represAsDataFrame.R
\name{represAsDataFrame}
\alias{represAsDataFrame}
\title{Coerce Representative list to a data.frame}
\usage{
represAsDataFrame(nrlist)
}
\arguments{
\item{nrlist}{The output of
\code{\link[=getLeontisList]{getLeontisList()}} or
\code{\link[=getAltRepres]{getAltRepres()}}.}
}
\value{
A data frame with the data of the representative structures
}
\description{
Takes the output of getLeontisList or getAltRepres, which
represent molecules with the format "XXXX|M|C+XXXX|M|C" (XXXX: PDB ID;
M: Model; C: Chain) and returns a data.frame with a more friendly
structure:\cr
Col 1: Equivalence Class.\cr
Col 2: PDB ID.\cr
Col 3: Model.\cr
Col 4: Chain.\cr
Columns 2 to 4 can be the direct input of
\code{\link[=pipeNucData]{pipeNucData()}}
}
\examples{
 data <- getLeontisList(release=3.2, threshold="1.5A")
 reps <- represAsDataFrame(nrlist=data)

}
\author{
Diego Gallego
}
