% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rvec_plot}
\alias{rvec_plot}
\title{Scatter plot into a png file}
\usage{
rvec_plot(ntID = NULL, df_rvectors, o = "", width = 15, height = 15,
  bg = "white", units = "cm", res = 200, cex = 0.6, cols = 3)
}
\description{
Given a data.frame with three columns generate a scatter plot
}
