% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CIF.R
\docType{methods}
\name{rVector}
\alias{rVector}
\alias{rVector,CIF-method}
\alias{rVector,character-method}
\alias{rVector,ANY-method}
\title{Compute the rVectors between the bases of a RNA structure}
\usage{
rVector(cif, pdb, outformat = "rvector", simple_out = T)

\S4method{rVector}{CIF}(cif, outformat = "rvector", simple_out = T)

\S4method{rVector}{character}(cif, outformat = "rvector", simple_out = T)

\S4method{rVector}{ANY}(pdb, outformat = "rvector", simple_out = T)
}
\arguments{
\item{cif}{A CIF object as otained from cifParser.}

\item{pdb}{A pdb object as obtained from cifAsPDB or read.cif/read.pdb
(from bio3d package).}

\item{outformat}{A string indicating the output format. This could be:
"rvector", "vector_coord" or "cylindrical_coord".\cr "rvector": (r(x)/a,
r(y)/a, r(z)/b), being a=5 and b=3. \cr
"vector_coord": (r(x), r(y), r(z)).\cr
"cylindrical_coord": (rho, phy, z).\cr
See reference paper for more details. This does not apply to the \code{gamma}
and \code{beta} angles.}

\item{simple_out}{A logical to simplify the output to a matrix.}
}
\value{
A list of data.frames for the values of each base or a single
matrix with all the data appended.
}
\description{
Given a RNA structure it computes the "r" vetors between all bases, as
defined by Bottaro et al. 2014 (The role of nucleobase interactions in RNA
structure and dynamics). This function is the basis to compute the epsilon
RMSD (see the same paper for details and run it with \code{\link[=eRMSD]{eRMSD()}}).\cr
Furthermore, it also computes two more metrics:\cr 1. The \code{gamma angle},
as obtained between the x axis of the coordinate system for all the bases
projected on the referece base plane and the x axis of the latter. This is
a metric of the relative rotation between bases along the orthogonal to the
base plane axis (z).\cr 2. The \code{beta angle}, as obtained between the base
planes normal to account for the degree of coplanarity between bases.
}
\examples{
cif <- cifParser("2d18")
model1 <- selectModel(cif=cif, model=1)
vectors <- rVector(cif=model1, simple_out=FALSE)

}
\author{
Diego Gallego & Leonardo Darre
}
