% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ntID_neighbours.R
\name{select_ntID_neighbours}
\alias{select_ntID_neighbours}
\title{Select neighboring nucleotides}
\usage{
select_ntID_neighbours(ntID, ntinfo, prev = 2, post = 2, info = "ntID",
  verbose = T)
}
\arguments{
\item{ntID}{an obejct of class vector with the desired nucleotide of
analysis.}

\item{ntinfo}{a data.frame with the data. It should contain at least the
columns "pdbID", "chain", "model", "resno", "insert" and "ntID" (as the
output of make_ntinfo() function.}

\item{prev}{Number of desired 5' neigbours to be returned.}

\item{post}{Number of desired 3' neigbours to be returned.}

\item{info}{Column name of the desired data to be returned.}

\item{verbose}{A logical to print details of the process.}
}
\value{
A vector with the desired data, extracted from the input data.frame
}
\description{
Given a ntID and a number of neighbors returns the ntIDs of the whole
polinulceotide. The interesting point of this function is that in the case
of asking for too much neighbors, a vector containing as many NA will be
returned, so the output vector will always have the desired length.
}
\author{
Diego Gallego
}
