% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_by_ntID.R
\name{trim_by_ntID}
\alias{trim_by_ntID}
\title{Calls trim_sphere to generate smaller pdb files}
\usage{
trim_by_ntID(cif = NULL, ntID, ntinfo, prev = 2, post = 2, verbose = T,
  file = NULL, ...)
}
\arguments{
\item{cif}{A cif/pdb object obtained from cifParser/read.pdb respectively
or a pdb ID so that the function can download the data. If NULL, the
function will extract the pdb ID from the ntinfo data frame (pdbID col).}

\item{ntID}{An integer/string with the desired nucleotide ID for
analysis.}

\item{ntinfo}{a data.frame with the data. It should contain at least the
columns "pdbID", "chain", "model", "resno", "insert" and "ntID" (as the
output of make_ntinfo function).}

\item{prev}{Number of desired 5' neigbours to be returned.}

\item{post}{Number of desired 3' neigbours to be returned.}

\item{verbose}{A logical to print details of the process.}

\item{file}{A string to save the output in a pdb formated file. If NULL the
fucntions just returns the pdb object.}

\item{...}{arguments to pass to trim_sphere (type ?trim_sphere for details)}
}
\value{
A smaller pdb object or a pdb file.
}
\description{
Given a data frame with nucleotide info (as obtained from make_ntinfo)
and the desired nucleotide index (ntID), the function returns a pdb
object or file allowing the user to select a number of 5' and 3' neighbors
in sequence and non-conected residues in a cutoff radius.
}
\author{
Diego Gallego
}
