% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_report.R
\name{validation_report}
\alias{validation_report}
\title{Download and read validation report for a PDB ID}
\usage{
validation_report(pdbID)
}
\description{
Download and read validation report for a PDB ID
}
