% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veriNA3d-package.R
\docType{package}
\name{veriNA3d}
\alias{veriNA3d}
\alias{veriNA3d-package}
\title{veriNA3d: Structural bioinformatics package for data mining of the PDB.}
\description{
The data pipeline starts obtaining, cleaning and exploring data. When
considering biological data, reproducibility, scalability and
standardization usually become issues. veriNA3d provides tools to analyse
PDB (Protein Data Bank) structural data - with a focus on RNA - while
making these issues easy to address. It makes data mining straightforward
by providing three categories of functions to get, manage and explore data.
}
\section{Get/query functions}{

The query functions ...
}

\section{Manage/compute functions}{

The manage functions ...
}

\section{Explore/save functions}{

The explore functions ...
}

\author{
Diego Gallego
\href{mailto:diego.gallego@irbbarcelona.org}{diego.gallego@irbbarcelona.org}
}
