% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cif_accessors}
\alias{cif_accessors}
\alias{cifEntry}
\alias{cifAudit_conform}
\alias{cifDatabase_2}
\alias{cifPdbx_database_status}
\alias{cifAudit_author}
\alias{cifEntity}
\alias{cifChem_comp}
\alias{cifExptl}
\alias{cifStruct}
\alias{cifStruct_keywords}
\alias{cifStruct_asym}
\alias{cifAtom_sites}
\alias{cifAtom_type}
\alias{cifAtom_site}
\title{Accessors to a CIF object}
\usage{
cifEntry(x)

cifAudit_conform(x)

cifDatabase_2(x)

cifPdbx_database_status(x)

cifAudit_author(x)

cifEntity(x)

cifChem_comp(x)

cifExptl(x)

cifStruct(x)

cifStruct_keywords(x)

cifStruct_asym(x)

cifAtom_sites(x)

cifAtom_type(x)

cifAtom_site(x)
}
\arguments{
\item{x}{a CIF object}
}
\value{
\itemize{
\item {cifEntry} \code{Character} with the mmCIF PDB ID
\item {cifAudit_conform} \code{Character} vector with dictionary version
\item {cifDatabase_2} \code{Data.frame} with cross-references
\item {cifPdbx_database_status} \code{Character} vector with deposition data
\item {cifAudit_author} \code{Data.frame} with author names
\item {cifEntity} \code{Data.frame} with molecules & ions in the structure
\item {cifChem_comp} \code{Data.frame} with residues records in the structure
\item {cifExptl} \code{Character} vector with experimental technique
\item {cifStruct} \code{Character} vector with author description of the structure
\item {cifStruct_keywords} \code{Character} vector with author selected key words
\item {cifStruct_asym} \code{Data.frame} with chain-entity equivalences
\item {cifAtom_sites} \code{Character} vector with details about the
crystallographic cell
\item {cifAtom_type} \code{Data.frame} with about the atoms in structure
\item {cifAtom_site} \code{Data.frame} with atomic coordinates
}
}
\description{
Accessors to a CIF object
}
\author{
Diego Gallego
}
