% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cifCheck}
\alias{cifCheck}
\title{Is an Object of Class CIF?}
\usage{
cifCheck(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A logical
}
\description{
Checks whether an object is of Class CIF.
}
\examples{
cif <- cifParser("1bau")
cifCheck(cif)

}
\author{
Diego Gallego
}
